+++ cpr.js
@@ -39,6 +39,7 @@ Hooks.once("init", async () => {
       "ammo",
       "gear",
       "skill",
+      "criticalInjury",
       "vehicle",
       "cyberdeck",
       "program",
@@ -76,7 +77,7 @@ Hooks.once("ready", () => {
   if (!game.user.isGM) return;
   // This defines the version of the Data Model for this release.  We should
   // only update this when the Data Model Changes.
-  const DATA_MODEL_VERSION = "0.73.1";
+  const DATA_MODEL_VERSION = "0.73.2";
   // Get the version of the data model being used for the loaded world. At
   // the end of a migration, this is updated with the current version of the
   // CPR system.
@@ -84,7 +85,7 @@ Hooks.once("ready", () => {
   // Determine if we need to perform a migration
   const needsMigration = dataModelVersion && isNewerVersion(DATA_MODEL_VERSION, dataModelVersion);
   if (!needsMigration) return;
-  Migration.migrateWorld(dataModelVersion);
+  Migration.migrateWorld(dataModelVersion, DATA_MODEL_VERSION);
 });
 
 registerHooks();